requirements.txt
streamlit
langchain
langchain-groq
langchain-community
python-dotenv
bs4
faiss-cpu
groq
PyPDF2
langchain-google-genai
pypdf
flask
werkzeug
joblib
flask-session
gunicorn